docker/fate_prediction-osx/requirements.txt
scdiffeq==v0.1.7
scdiffeq-analyses==v0.0.1rc0
LARRY-dataset==v0.1.0dev5
wandb